chore: move UI files to the presentation package
