+++ app/src/main/AndroidManifest.xml
@@ -14,7 +14,7 @@
         android:theme="@style/Theme.Blocky"
         tools:targetApi="31">
         <activity
-            android:name=".MainActivity"
+            android:name=".presentation.screen.MainActivity"
             android:exported="true"
             android:label="@string/app_name"
             android:theme="@style/Theme.Blocky">
