+++ app/src/main/java/com/github/jvsena42/blocky/presentation/ui/theme/Type.kt
@@ -1,4 +1,4 @@
-package com.github.jvsena42.blocky.ui.theme
+package com.github.jvsena42.blocky.presentation.ui.theme
 
 import androidx.compose.material3.Typography
 import androidx.compose.ui.text.TextStyle
